* commit 'c755c54fac1085bd4e57159158252371d2f1c7f0':
  Fixing warning in 32 bit build

+++ src/core/lib/core_data_model/collection_model.cpp
@@ -107,7 +107,7 @@ Collection & CollectionModel::getSource()
 
 AbstractItem * CollectionModel::item(int index) const
 {
-	if (items_.size() <= index)
+	if (items_.size() <= (size_t)index) 
 	{
 		items_.resize(index + 1);
 	}
